(htparse): Consistently allow LF only where CRLF is required.
We're expected to be used by lots of broken clients relying on the
server to be forgiving, so just do like the major HTTP servers and
allow LF only as line terminator.

The standard behaviour of htparse to allow LF in most of the places in
the parser and return error in a few places doesn't make sense. And in
two places the state machine would exit without setting eny error,
parsing the input buffer beyond the LF, or advancing the state.